handled review comments
